// src/models/codicefiscale.model.ts
export interface CodiceFiscaleForm {
  name: string;
  lastname: string;
  gender: string;
  city: string;
  dayDate: string;
  monthDate: string;
  yearDate: string;
}

export interface CodiceFiscalePartials {
  name: string;
  lastname: string;
  monthDate: string;
  yearDate: string;
  dayGender: string;
  city: string;
  controlCode: string;
}